requirements.txt
# === Librerías base para ETL ===
pandas>=2.0.0
pyarrow>=14.0.0
numpy>=1.24.0

# === Cargar variables de entorno ===
python-dotenv>=1.0.0

# === PostgreSQL / Load ===
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
